backend/requirements.txt
# Django ve REST framework
Django>=4.0,<5.0
djangorestframework
psycopg2-binary
django-cors-headers
dotenv
requests
polyline
decimal
geojson
geopy
python-dateutil
Pillow

# Veritabanı
dj-database-url==2.3.0

# Servis ve sunucu
gunicorn==23.0.0
whitenoise==6.9.0

# Zamanlanmış görevler
django-crontab==0.7.1
python-crontab==3.2.0

# Web scraping
selenium==4.29.0
beautifulsoup4==4.13.3
pandas==2.2.3
Unidecode==1.3.8
django-celery-beat==2.7.0

# E-posta ve şifre sıfırlama araçları
django-templated-mail==1.1.1
pyjwt==2.9.0

# Ortam değişkenleri ve yardımcı kütüphaneler
urllib3==2.3.0
numpy==2.2.4
pytz==2025.1
uuid==1.30

# Güvenlik
django-password-validators==1.7.3

# Ek bağımlılıklar
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
sqlparse==0.5.3
tzdata==2025.1

# Debug ve Test
django-debug-toolbar==4.3.0